+++ src/caffe/layers/base_conv_layer.cpp
@@ -314,7 +314,7 @@ void BaseConvolutionLayer<Dtype>::Reshape(const vector<Blob<Dtype>*>& bottom,
   // LOG(ERROR) << "final thread number: " << num_of_threads_;
 #endif
 
-  int col_buffer_mt_size = num_of_threads_ * col_buffer_.count();
+  long col_buffer_mt_size = num_of_threads_ * col_buffer_.count();
   int weight_diff_mt_size = num_of_threads_ * this->blobs_[0]->count();
 
   col_buffer_mt_.resize(col_buffer_mt_size);
